requirements.txt
gitpython
vdf